Final working optimization algorithm
This algorithm will choose the most efficient direction to move around a circle to replicate the desired angle given two inputs. This allows for better and smoother swerve drive because it can travel around the circle past the sensor transitions and can reverse the motor direction so that the max distance moved is only 90 degrees unless spinning while moving forward in a "tornado" maneuver.